Prevent bubble replication lag bug by reading from master on failure
Summary:
We've been seeing some errors on ASIC [1] that say `bubble X does not exist, or has expired`. Since we don't expire bubbles yet, and we only call this endpoint after having actually created the bubble, I conclude this is likely caused due to replication lag.

To fix that, and at the same time avoid always querying the master SQL replica, which is undesirable, let's first try a normal read, and only if it yields no results, then we query master. This should only add an overhead when replication lag happens, which is quite rarely.

Also, we don't use this `open_bubble` function to check if a bubble exists, we always call it when we know it already exists, so this should not add any undesirable overheads in that case.

Reviewed By: mrkmndz

Differential Revision: D34045137

fbshipit-source-id: 5928f7d457d5104505df847d91c83eee0582e343