Make integration test less flaky and more insightful
Summary:
`test-snapshot-update.t` was identified as flaky.

I modified `assert_on_base_snapshot` to print more information on failure, and stress tested it with:
```
> buck2 test eden/mononoke/tests/integration/facebook:snapshot -- --stress-runs 50 test-snapshot-update.t
```

On 2 of those 50 runs, it failed because `hg status` didn't realise the `modified_file` was modified. Notice that we also `cat` that file and check its contents, and the content was indeed correctly modified, so this is likely a problem in `hg status` somehow being out of sync.

Since checking `hg status` was not that important in this case, I just removed it to make the test less flaky, and we still should get good signal.

Reviewed By: HarveyHunt

Differential Revision: D34110526

fbshipit-source-id: fe726e0652db202b7668ff666e30df3bff453aab