nativecheckout: don't remove files that didn't exist before
Summary:
When doing a clone, the nativecheckout action plan included a remove
action for every file not in the sparse checkout. This is because it didn't
account for the file not existing in the old manifest and just blindly queued
the remove.

On Linux this is pretty fast, but on Windows this is extremely slow to do for
every path. It's probably slow because of NTFS, but also because we do some path
auditing logic for Windows and Mac.

This fixes it to not do remove for a file that wasn't in the old manifest.

Reviewed By: sggutier

Differential Revision: D34188271

fbshipit-source-id: 714b2dcf98cfa9ebec287f00d7cd5397126f693e