blobrepo: use bonsai hg mapping method only
Summary:
There are currently multiple ways to map between bonsai and hg changesets: using the bonsai-hg mapping repo attribute, or using the method on blobrepo that is very similar (but takes `ctx` by value).  Standardize on using the repo attribute, and convert all callers.

There is one exception to this: when callers wish to trigger derivation, they
continue to call `BlobRepoHg::get_hg_bonsai_mapping`.  Despite its name, this
method triggers derivation in the bonsai-to-hg direction only.  We will
separate out this functionality in the future.

Reviewed By: mitrandir77

Differential Revision: D34071889

fbshipit-source-id: f0c4ce7987376702dc36eb27683d1fcb8a22df60