bonsai_hg_mapping: store repo id and make interface consistent
Summary:
The various mappings are inconsistent in how they work.  Some, like the
bonsai-hg mapping, require the client to pass the repo_id on every request (and
behaviour is undefined if the client gets this wrong).  Start to make them
consistent and clean up this gap, starting with the bonsai-hg mapping.

Reviewed By: mitrandir77

Differential Revision: D34059165

fbshipit-source-id: 12c62e0e5f08685929218088bad11a55305795df