rebuild inMemory SC on the server with the heads provided by client during the pull
Summary:
When the master bookmark moved backwards, clients may have old location of the bookmark pulled locally. It was failing due to the fact that server couldn't find that bookmark location in it's IdMap.
Let's allow SC to build idMap in memory up to the client heads so it can correctly respond to the pull request.
We will only check some number of parent commits from the client head to check if they are ancesstors of master to avoid expensive SC rebuilds. And we still need to check for the merge commits with large number of commits not in the SC. (D33654925 (https://github.com/facebookexperimental/eden/commit/78c19c7a79be90c6b6b8a80048fd8a2c14f3d24d) already reduce the possibility of this)

Reviewed By: farnz

Differential Revision: D34041066

fbshipit-source-id: aa7274f97c675e32e87a791aca336061dc75d645