bonsai_svnrev_mapping: store repo_id and make interface consistent
Summary:
The various mappings are inconsistent in how they work.  The bonsai-svnrev
mapping has a separate struct to hold the repo_id, and must be constructed and
used in a different way to the others.  Continue to make them consistent and
clean up this gap, this time with the bonsai-svnrev mapping.

Reviewed By: yancouto

Differential Revision: D34140241

fbshipit-source-id: 432f1394b05431625349781b95430b37a0478c6c