bonsai_globalrev_mapping: store repo id and make interface consistent
Summary:
The various mappings are inconsistent in how they work.  Some, like the
bonsai-globalrev mapping, require the client to pass the repo_id on every
request (and behaviour is undefined if the client gets this wrong).  Continue to
make them consistent and clean up this gap, this time with the bonsai-globalrev
mapping.

Reviewed By: yancouto

Differential Revision: D34093084

fbshipit-source-id: de67d4da54039a35baa4db65d208d71f1195975c