bonsai_git_mapping: make builder interface consistent
Summary:
The various mappings are inconsistent in how they work.  The bonsai-git
mapping is mostly where we want it to be, however the builder name is wrong,
and the implmentation is mixed into the same module as the trait.

Clean this up by spliting the implementation out to its own module, and
renaming the builder to `...Builder`.

Reviewed By: yancouto

Differential Revision: D34196512

fbshipit-source-id: e8de38d4d495ac911a6fa1e1c65c10337c5bf281