cacheblob: close the persist hole in MemWritesBlobstore
Summary:
As described in D32638849 (https://github.com/facebookexperimental/eden/commit/ec370e31b74b284c8d7255177b17bcdf06972309), there is a footgun in `MemWritesBlobstore::persist`.
For the duration of the persist, the items being persisted are temporarily
invisible from outside.

Previous attempts to fix this were made by trying to stop people from accessing
the blobstore while the persist was ongoing.  Instead, let's just extend
`MemWritesBlobstore` so it doesn't have the hole in the first place.  While we
are persisting the items, we keep a reference to the items being persisted in
the cache.  Any key lookups that fail for the live cache can also be checked
against the flushing cache.  Once the flush has completed, we can clear those
items out.

A downside here is that we must clone the keys and values: there need to be two
copies, one left in the flushing portion of the cache, and one that is passed
by value to `Blobstore::put`.  However, since this is a `String` and a
`BlobstoreBytes` (which reference-counts the bytes), cloning should be
relatively cheap.

Reviewed By: mitrandir77

Differential Revision: D34074710

fbshipit-source-id: 4562833266b80bf95ed21a01b4f871e341b2f357