drawdag: add string sucessor utility
Summary:
This will be used in the next change to create a range of linear commits
easily, like A..=Z, A01..=A20, etc.

The implementation is inspired by the `String::succ` method in the Ruby
programming language, except for the non-alphanumeric case, because that's
a case not intersted by drawdag, and finding the next valid utf-8 code point
in Rust seems non-trivial.

Differential Revision: D34163017

fbshipit-source-id: 9d1036cabab7a7f33e93a00c3e7c90e40534c994