merge: remove checkunknown content check
Summary:
checkunknown is a large source of random slowdowns for users, since if
they have a lot of untracked files in their working copy, it could result in
downloading a lot of data, reading the files off disk, and comparing it. This
causes a lot of pain, and in contrast the checkunknown functionality likely
saves users time very rarely (vs just rewriting the files on disk), so let's
remove it.

Reviewed By: quark-zju

Differential Revision: D34113766

fbshipit-source-id: 442b347e5484422c5b27154763bd878b2bb8803e