mononoke: Remove readonly check from unbundle path
Summary:
A Mononoke repo can be marked as read only in config, or be read only
depending on a database value. The only place that this used to be checked was
at the start of the unbundle code.

In order to allow sync jobs and commit cloud to write bundles to Mononoke, a
`BYPASS_READONLY` pushvar was checked in the unbundle code. Both of these
clients hard code the bypass push var.

After some refactoring, we now check whether a repo is read only both at the
start of the unbundle path and also when a bookmark is moved.

This diff removes the read only check at the start of the unbundle path. The
motivation is so that usage of the `BYPASS_READONLY` pushvar can be restricted
to just sync jobs. Once this diff lands, I'll be able to remove the
`BYPASS_READONLY` pushvar from the commit cloud code path.

Before this change, a client trying to unbundle to a locked repo will be given
an error before they start to upload content. After this change, they will only
discover the repo is locked after attempting to move a bookmark.

Reviewed By: markbt

Differential Revision: D34108102

fbshipit-source-id: 8b5cf5ddded6c699686592141bc7664c1ed15924