rename clap to clap-old in most of mononoke
Summary:
The autocargo was complaining about collision between clap and clap-old
when we tried to migrate target that shared Cargo.toml file with unmigrated
target. This should be enough to alleviate problem.

NOTE: I've also had to make a copy of value_t macro from clap as it was
referring the crate by name.

Reviewed By: markbt

Differential Revision: D34219408

fbshipit-source-id: b0bb1328d8808e7387944d8df14dc374e569ece2