Auto-derrive `quickcheck::Arbitrary` for amenable types in `edenapi/types`
Summary:
Use the `quickcheck_arbitrary_derive` crate to derive the `Arbitrary` trait for all the types for which the auto-derived impl would match the existing impl.

note: I did not commit the updated snapshots from the `buck test ... -- --env INSTA_UPDATE=1` execution because I still don't know if they are correct or not. :) yelp! :D

Reviewed By: yancouto

Differential Revision: D34161995

fbshipit-source-id: 567c6c43b1fb666abd460c4a7ac62bf095220600