mononoke_api: update inner repo's blobrepo when opening bubbles
Summary:
When a bubble is opened, `RepoContext` has two `BlobRepos`: the one inside the
context, which has the bubble opened, and the one inside the `InnerRepo`, which
does not, and is shared with the main repo.

This could be a problem in the future when we start using `InnerRepo` instead
of `BlobRepo`.  Change it so there is only one `BlobRepo`: the one inside the
`InnerRepo`.

Reviewed By: yancouto

Differential Revision: D34222496

fbshipit-source-id: e576fd72b8f8c9187ed9a518ce59d30e189afae2