physicalfilesystem: propagate 'rustwalkerthreads' config to pendingchanges
Summary:
This diff passes the `workingcopy.rustwalkerthreads` config value along to the physical filesystem `pending_changes` function, allowing it to use a multi-threaded walker when `rustwalkerthreads > 0`.

We also need to use `extract_matcher` to get a Rust Matcher from a Python one for use in the multi-threaded walker. Using `UnsafePythonMatcher` causes the walker to deadlock (presumably due to the GIL -- see P480409358 for an example backtrace of a hung thread). This was the last usage of `UnsafePythonMatcher`, so I removed it as well.

Reviewed By: DurhamG

Differential Revision: D34251644

fbshipit-source-id: 1f11f2038ebcc6e098432132b9e5726b25c18911