bookmarks_movement: check whether bookmarks are being redirected
Summary:
Currently the source control service will accept bookmark moves for repos that
are the small-repo in a push-redirected megarepo config.  This means it's
possible for users to get bookmarks out of sync with a bookmark move, which
requires manual intervention to fix.

Prevent this from happening by adding a check in `bookmarks_movement`.  If a
repo has push-redirection enabled, then bookmarks of that kind are not allowed
to be moved by the source control service.  Users should move the bookmark in
the source-of-truth repo and allow the backsyncer to move it in the
push-redirected repo.

Reviewed By: mitrandir77

Differential Revision: D34222495

fbshipit-source-id: 64a0fa929af53c9991c2a0f8bb84a21e4a1de9fc