Move copy from BlobstoreWithLink to Blobstore trait
Summary:
**Context**: See T110025458

This diff adds the `copy` method to the blobstore trait. It is a very basic method that copies one key to another. The default implementation simply loads the value into memory and stores it again, but faster implementations can exist.

I moved the implementations from `BlobstoreWithLink`, and renamed that to `BlobstoreUnlinkOps`, as we don't need this on the default Blobstore trait (for now).

Reviewed By: farnz

Differential Revision: D33193940

fbshipit-source-id: b10beecbd827396d1685f8196bab290e81ba4039