Copy file to another bubble using as little memory as possible.
Summary:
**Context**: See T110025458

This diff is the final piece to solve T110025458. It uses the `filestore::copy` method added on D33193941 in conjunction with `copy` added on D33193941 to have fast copying.

Testing locally, the improvement is a bit disappointing. For a 1G file, it went from 18s to 12s. I expected much, much less. That said, this was a prod build vs a non-opt build with a lot of print-debugging.

Reviewed By: farnz

Differential Revision: D33193939

fbshipit-source-id: 3c976f259f66d27da1f19c9d129ffdbb64745a8e