dag: extract setdiscory algorithm to a separate function
Summary:
The `calculate_definitely_unassigned_vertexes` function is an important
optimization to avoid excessive remote lookups when adding new commits
to the lazy graph. It contains the a "discovery" algorithm that uses a
bisect-ish way to find out what vertexes are known or unknown remotely.

That "discovery" algorithm can also be useful in other places, including:
- Figure out what vertexes need to be "reassigned" from the non-master
  group to the master group.
- With proper Python bindings, replace the Python `setdiscovery.py`
  entirely.

This diff makes the above possible by moving the algorithm to a separate
function.

Differential Revision: D33865775

fbshipit-source-id: 3b54de8929dd69ea80bdd2d8667b1791fe60ed50