scs_server: implement commit_path_last_changed
Summary:
Implement a new method pair for the Source Control Service:
`commit_path_last_changed` and `commit_multiple_path_last_changed`.

These methods return the last commit that changed a path for one or more paths
in a commit.  If something exists at that path, this will be the last commit to
modify a file at that path or a file in a directory at that path.  If nothing
exists at that path now, but something previously existed, then this will be
the commit that deleted the file or directory at that path.

Similar to the other `commit_multiple_path` method,
`commit_multiple_path_last_changed` does not return entries for paths that never
existed.  This is done because efficient manifest traversal prunes these paths
before they are reached.

Reviewed By: yancouto

Differential Revision: D34047425

fbshipit-source-id: 4290e3193aa7834e6d44c90eb4b2626d580b43ae