fix stale inode unload crash
Summary:
In forgetStaleInodes we are checking for unlinked unloaded inodes and
de-referencing them. The dereferencing is happening inline with where we
are iterating through unloaded inodes.

This is causing us to invalidate an iterator that we are using. Which causes
a use after free.

let's collect the inode number of inodes to de-reference and de-reference them
after we finish iterating.

Reviewed By: xavierd

Differential Revision: D34315130

fbshipit-source-id: b390959bd26edf1545e900f0e5f72fc1885efa3a