show stack traces of edenfs process in `eden rage`
Summary:
Prints the output of `sample $edenfs_process_id` in `eden rage`.

On linux we can use `ps f` to see the process hierarchy we want to have similar
information on macOS. We would also like to have info on the state of the
threads, since we can not log into their machine and look for our selves this
would be a good thing to have in the rage.  This give us visibility into eden to
see where the problem might be when there is a rage particularly if/where eden
is hanging.

Reviewed By: xavierd

Differential Revision: D16745028

fbshipit-source-id: 8dcbc1d80b45c3ebe4d065f445f56800104dbd25