Extract scribe log logic from create_changeset
Summary:
This does some changes to scribe logging logic:

Move it from `bookmarks_movement` to `blobrepo`. Being in `bookmarks_movement` was wrong, as it doesn't only happen in that case. `blobrepo` is not ideal but it's better, as it's close to where we save changesets, which is more or less when this happens.

This is done so the next diff can reuse more of the logging logic without having to duplicate code.

Reviewed By: markbt

Differential Revision: D34145742

fbshipit-source-id: 1dfb94775ea60e4b83412ae80f5c6d3f5e65d952