Support logging commit cloud commits uploaded via edenapi to scribe
Summary:
markbt noticed some commit cloud commits weren't being logged to the scribe category. This is because the EdenApi path to create changesets doesn't do the logging.

This diff makes sure we also do the logging in that case.

Ideally, we'd have an unified place to create changesets, be it from hg changesets, git, or from raw data, but that needs a lot more refactoring and it's riskier.

Reviewed By: markbt

Differential Revision: D34146757

fbshipit-source-id: e82f14e1be6c598f89722b68c20cd6fc572633e0