snapshot: do not print commit cloud output on plain mode
Summary:
Right now, when we create a snapshot on top of a commit that's not on commit cloud, we output things like `queue 0 files for upload` even when we are on plain mode, in which we'd like to only output the hash of the snapshot.

So I made it so the ui is quiet when on plain mode, which makes it not print that.

Reviewed By: quark-zju

Differential Revision: D34378357

fbshipit-source-id: 69b17147b8e72c6d72176f6069d729689130fdc9