physicalfilesystem: fix incorrect logic leading to duplicate path lookups
Summary:
This diff fixes some incorrect logic in the physicalfilesystem pendingchanges Rust code that leads to duplicate entries in the "lookup" list (files that need to have their contents checked to know whether they're dirty). We need to fix this, otherwise we'll do 2x the number of (potentially expensive) checks, and we'll need to deduplicate later anyway.

I believe this bug was originally introduced in D25697099 (https://github.com/facebookexperimental/eden/commit/02606da6c5e2697712014fc571763f32a9b8304e) -- the change to this region of the code changes the behavior. I believe this was just an oversight when the code was refactored, since the change doesn't have anything to do with the purpose of the diff.

Reviewed By: DurhamG

Differential Revision: D34403900

fbshipit-source-id: 4c38e6a0ba46dfefc8f7ba327b943cb7ccee7fc5