scmstore: inject an "unknown fetch error" in scmstore
Summary:
We are seeing fetch failures with empty error lists ("[]"). This should
never happen.  Every error should be attributable to an edenapi or a lfs server
fetch failure. There is a clear code path in scmstore that adds an empty error
list if we don't have an error. Let's make this explicitly call out that it's an
unknown error, so we can track these errors down in the future.

Reviewed By: quark-zju

Differential Revision: D34254835

fbshipit-source-id: baa46d641d9b951b1563ac643517dcef15efa4bc