nativecheckout: only filter actions once
Summary:
When resuming a checkout we need to go over all the actions and check
if they match the contents on disk. In the old code we did this twice, once when
checking unknown files and once when doing the actual checkout. When the number
of files to check was large, like when resuming a large clone, this was
extremely slow (i.e. many minutes).

Let's only do it once.

Differential Revision: D34314688

fbshipit-source-id: 848dc4e76cb72f173741a9061bd8c71d7120a4bc