vfs: read symlink link instead of symlink target content
Summary:
The nativecheckout resume feature was failing when a symlink had been
written because the "is it up-to-date" check compared the expected link path
with the contents of the file at the link target.

Let's change vfs::read() to return the symlink link instead of the content, as
that is what Mercurial wants. I can't think of any case where Mercurial ever
wants to read through to the symlink target.

This does add the overhead of a stat call to each read.

Differential Revision: D34314987

fbshipit-source-id: 676647a16f65fa516d6ef97e21d137a1ae137b3f