scmstore: annotate lfs other errors as invalid
Summary:
LFS fetch errors are reported via sha256, and in theory we should have
a Key for every sha256, so we can log keyed rrors all the time. Let's annotate
the non-keyed errors as invalid.

Eventually we'll come back and remove other errors entirely, once we can confirm
this is never hit.

Differential Revision: D34256994

fbshipit-source-id: b00db226d8ef1f222a345f0388c3cb2a5773e008