sparse: fix debugsparsematch to not always union v2 profiles
Summary: Version 2 sparse profiles get unioned at the top level (e.g. if you do "hg sparse enable A" and "hg sparse enable B" and they are v2 profiles, you get "matches(A) OR matches(B)"). However, debugsparsematch was always applying the union logic to the given profile, so for example "hg debugsparsematch -s A ..." would union profiles included in A, which does not match the actual behavior when you enable A.

Reviewed By: quark-zju

Differential Revision: D34463674

fbshipit-source-id: b5e5dfcdf2bd36e0e00c6a1b814fafbc6a6d59cc