edenapi: log unfetchable blobs and trees with EdenAPI
Summary:
Add EdenAPI fallover scuba logging to `HgBackingStore`.

The reason that this is being added to `HgBackingStore` instead of `HgQueuedBackingStore` is that we don't have good access to the actual point where trees and blobs are imported from HgImporter in there. As a result we have to get StructuredLogger in there and send events directly from there.

Reviewed By: xavierd

Differential Revision: D34217906

fbshipit-source-id: 795b61ef6a2ea52e8b67cd08633ec017e5bdbb06