refactor ServerState constructor
Summary:
Our aim is to construct as few things as possible in the ServerState constructor. Instead, we should preconstruct these objects, create shared pointers, and pass those into the ServerState constructor. This will allow us to pass in "null" pointers to ServerState for testing purposes in the future.

In this diff, we preconstruct a shared_ptr<ReloadableConfig> inside EdenServer so that we can pass it to the ServerState constructor.

Reviewed By: chadaustin

Differential Revision: D34541321

fbshipit-source-id: 99465fe08099ed95179e56b2e39a97ab68112733