commitcloud: always check metalog root for changes
Summary:
Multiple transactions can happen within a command.  If the last transaction
doesn't change the metalog, then the commit cloud check for whether anything
changed won't see a change, and the background backup won't be started.

Instead, ignore whether there was a transaction at all, and just check to see
if the metalog has a new root after the command has completed.  If it has,
start the background backup.

Reviewed By: DurhamG

Differential Revision: D34556502

fbshipit-source-id: d70c86075a4a4a311de67f33b80f5f526b978411