Create DeletedManifestCommon trait and use it for deleted manifest
Summary:
This is a **refactor** that creates a trait called `DeletedManifestCommon`, implements that for `DeletedManifest`, and then only uses the trait on the derive functions for deleted_manifest.

It is a first step towards deleted_manifest_v2, as we will try to share as much code as possible with v1, and reuse the main logic for derivation.

High level idea of this first part:
1. Stop depending on DeletedManifest in derivation and instead depend on a trait **(this diff)**
2. Modify trait and derivation so it doesn't need to iterate over all subentries when possible.
3. Implement base DFv2 structures.
4. Implement trait for DFv2 structures and reuse derivation code.

Reviewed By: markbt

Differential Revision: D34551121

fbshipit-source-id: 411463dc7c87bef949d9f9e7ce5834eab6de1c42