On DF derivation, try to minimise the number of "list" calls.
Summary:
Context: D34551121 (https://github.com/facebookexperimental/eden/commit/6629440ff7b6d4223fa4b8232f5a5aa45051ba6b)

This diff is a refactor of deleted manifest derivation, which shouldn't change correctness or functionality, but just build things in a different way that opens up space for optimisation in the future.

It changes current derivation by, when deleting a given deleted manifest node, instead of looking at **all** its children, trying to derive them (and possibly just copying the same id), we instead only look at the children that were modified on the current changeset (if doing a merge, we might still need to look at everything).

So, in terms of the recently implemented `DeletedManifestCommon` trait, instead of using a `::new` method to create an object, which takes a list of all its subentries, we instead use a new `::copy_and_add_subentries` method, which "copies" all subentries from the given node (the parent of the current node), and then just adds and removes a few other entries, which size should be bounded by the number of changes in the commit.

This will allow faster derivation in the future, when using sharded maps. Notice that for now, this doesn't really improve efficiency at all, as we **always load everything in memory** in deleted manifest v1.

When considering the steps on D34551121 (https://github.com/facebookexperimental/eden/commit/6629440ff7b6d4223fa4b8232f5a5aa45051ba6b), this is part of step 2.

Reviewed By: markbt

Differential Revision: D34556462

fbshipit-source-id: 1c502ac0756eed270ef9b330f1c9284a775cc43f