change color of hg rage paste link for Windows
Summary:
by default, Windows PowerShell uses a blue background. I'm sure many users don't change this (but maybe they do and I'm weird).

The text is extremely hard to read if this is the case. We should update it to yellow on Windows (I highly doubt the yellow would clash with the typical PowerShell background).

Reviewed By: fanzeyi

Differential Revision: D34652357

fbshipit-source-id: b55d1912135b957629ffa027562e4a94db62e88f