Make async DeletedManifestCommon methods take blobstore
Summary:
Let's change `DeletedManifestCommon` so that `lookup` and `into_subentries` also take a blobstore and a context. For DMv1, that is not really necessary, but for v2 it will be, as they might need to fetch from the blobstore.

This also changes `lookup` to return an owned id, instead of a reference.

Reviewed By: mitrandir77

Differential Revision: D34617038

fbshipit-source-id: fe0601c51585ee4125cb94a2e35edc34910519f2