prjfs: ignore ERROR_PATH_NOT_FOUND on invalidation
Summary:
During an `hg update` Mercurial complained about the following error:
  error updating some/path: class std::system_error: Couldn't add a placeholder for some/path: 0x80070003: Error (0x80070003) The system cannot find the path specified.

While this didn't stop the update operation, this could be misinterpreted by
users leading to an increase in oncall load. The root cause of this issue is
that EdenFS has loaded TreeInodes that aren't present on disk, and would thus
try to invalidate them.

Reviewed By: chadaustin

Differential Revision: D34464274

fbshipit-source-id: 9ee54c4ab7475a8be5891325d3d1b7151a0878ed