hg: Remove BYPASS_READONLY pushvar from commit cloud upload path
Summary:
Before Mononoke was the source of truth for hg repos, D13804545 (https://github.com/facebookexperimental/eden/commit/5ff6db64eff92f8d72475f2656a1059a321fd296)
updated the hg client to pass a pushvar that allowed commits to be accepted to
a read only repo.

Since then, a couple of things have changed:
- Mononoke is the source of truth for all repos
- The `BYPASS_READONLY` pushvar is only checked when a bookmark move happens
  (D34108102 (https://github.com/facebookexperimental/eden/commit/71f58e3133f61a0ba3b71b58c378f605cf997adf))

This means that providing the pushvar is no longer necessary. Remove it from
the code.

In the future, I want to lock down use of the `BYPASS_READONLY` pushvar so that
only authorised users can use it. The first step of that is to make sure it's
only being passed when needed.

Reviewed By: farnz

Differential Revision: D34109906

fbshipit-source-id: 71c188d2322ab8cfecbe5aa7266654c2b56a4491