Don't swallow error on DM derivation
Summary:
The derive function used a `oneshot` channel to send back the response, however, this was not necessary, as we could just await the JoinHandle from the task it wanted the response from.

From a quick look in blame, it seems this is leftover from old to new futures migration, when doing this was probably harder.

This caused this function to sometimes swallow the real error and instead output "oneshot canceled" error, which is not great to debug.

Reviewed By: farnz

Differential Revision: D34720661

fbshipit-source-id: f3b936ffa6065a5bc927cbfb265a0ae7a5b51330