snapshots: Use VFS symlink and executable handling
Summary:
VFS already has logic to write symlinks and executables, let's depend on those instead of building our own, which means less surface for bugs.

We still need to fixup the executable permissions for the stuff that we "optimised out", but let's do that in Rust with VFS instead of in python.

This gets rid of some weird issue I've seen where it fails to restore a snapshot, but wasn't able to pinpoint the cause.

Reviewed By: mitrandir77

Differential Revision: D34643939

fbshipit-source-id: acab6f9f33e19237d95861087ba57d610731f8ab