edenapi: avoid retries for "cert required" errors
Summary:
We should not retry when there are missing or expired certs.
Unfortunately we were, because we get a generic TLS RecvError, so we couldn't
distinguish between a connection failure and a cert issue. This diff adds some
hacky error-message based heuristic to determine what kind of TLS error we got,
so we can retry on some but not others.

I also got rid of the existing ssl_in_message condition, since it seemed to be a
no-op.

Reviewed By: kulshrax

Differential Revision: D34734592

fbshipit-source-id: a71b90d15784b39e7dc61dad460aacf97187f63e