do not report conflicts if a file's object ID has changed but the contents have not
Summary:
When migrating to the new hg:object-id-format hash schemes, EdenFS
would sometimes report spurious conflicts from `hg update`. The issue
is that checkout treated the object ID as a proxy for equivalence,
which is not generally true. In that case, let's actually load the
inode in order to compare SHA-1s.

Reviewed By: xavierd

Differential Revision: D34742471

fbshipit-source-id: 230eed0d2d3e45315481a97e32deb41aa040d445