minitop: add a small Cursor abstraction for handling new lines
Summary:
There is a small bug in minitop where the cursor will never move from its
current row due to using scroll directive exclusively. In order to go to the
next row, the cursor needs to be moved down. To simplify the code, a Cursor
abstraction is added which detects when the end of the terminal is reached and
decides whether to scroll, or to go to the next line.

Reviewed By: genevievehelsel

Differential Revision: D34600897

fbshipit-source-id: 480ea919af93a8086def368dbf84afe7672f1a87