status: add missing FileStatus variants
Summary: When the Rust `Status` struct was originally created, it had 5 different `FileStatus` variants. However, `hg` actually considers two more: `ignored` and `clean` (although these are only revealed in `hg status` if the user explicitly requests them). In preparation for a Rust-native status command, this diff adds the missing variants. It also adds a bit of documentation to the `FileStatus` enum explaining (in what I believe to be clear terms) what each of them mean.

Reviewed By: DurhamG

Differential Revision: D34619612

fbshipit-source-id: fa07e9aab78e51a5aecf19f77bb5c5310e87df8a