cli: allow edenfsctl to show tracekback when configured
Summary:
Sometimes on Windows, we receive user reports that `edenfsctl` hangs indefinitely. We would like to see where it hangs in these cases. Adding this will allow us to see the traceback when ctrl-c edenfsctl with the environment variable set.

Note `--debug` is shared with Rust edenfsctl (where it enables verbose logging).

Reviewed By: xavierd

Differential Revision: D34731381

fbshipit-source-id: 24477d05c01d353f9f7a45fc3c5d28f29b62e3be