ui: Support number choices >= 10
Summary:
`choicetuple` currently assumes that the choice will be one character.
In cases where choice is a number >= 10, we see this behavior: {P485111360}
as the entered choice will never match the list of accepted choices.
The proposed fix is to check whether the non-`&` choice is a number
string, if so use that as the choice.

Reviewed By: yancouto

Differential Revision: D34705485

fbshipit-source-id: c538ae19b9a75077b63808f0574f7ac22fef7b5b