workingcopy: stub out Rust status
Summary:
This diff adds the Python/Rust glue necessary to start implementing the status command in Rust. Specifically, it allows swapping out the implementation of the "dirstate status" with a Rust function, by enabling the `workingcopy.ruststatus` config key.

The status command actually has a few different pieces to it. The full status command can take two arbitrary revisions (or a revision and the "working copy") and produce a status-formatted diff between the two. The first piece (stubbed out here) computes the status between the current revision (`.`) and the working copy.

The end goal is to have the full status functionality implemented in Rust. I chose to start here (by replacing the "dirstate status") because it 1) is something that needs to be implemented anyway and 2) allows us to test it with or without watchman (the pending changes from either concrete implementation are used), even though only the without-watchman case is implemented in Rust. Eventually, the pending changes logic will be called from Rust, rather than called in Python (potentially calling back into Rust) and transferred over to Python. An EdenFS backed repo will not work (yet) -- it bypasses this logic entirely.

In the mean time, to demonstrate that the glue works, I've implemented a basic, incorrect `status` implementation in Rust. Any files modified (or added) on the filesystem are treated as "modified", and any files deleted are treated as "removed".

Reviewed By: DurhamG

Differential Revision: D34623557

fbshipit-source-id: c16e15f5c6596c494a5a385da0ed244d7f27207e