commitcloud: make foreground syncs not best-effort
Summary:
If a user runs 'hg cloud sync' manually, they want to guarantee things
are backed up. In that case, we shouldn't fail if a background process mucked
with the repo in the meantime. Let's take the lock to ensure nothing mucks with
the repo while the sync is running in that case.

Reviewed By: markbt

Differential Revision: D34797187

fbshipit-source-id: 1732368a9565f12f49c395480341f19825b926f2